Correct bug in Mattermost not sending interactive dialog.
Mattermost expects to receive a json response to the
Interactive message.  Docs from mattermost show
that you *should* respond with a message
but in testing, the empty json gets accepted.  Responding with a message would impact every
command or interaction with mattermost, which
would be very annoying for the user.