Change CommandLog.params to JSONField.
This creates a data migration to convert the existing
data in params to the new JSONField.  Determines
if the database vendor is postgres, then it will migrate
to the new field. If the vendor is not postgres, then
there would be no data in the old field to migrate,
so these steps are skipped.